+++ packages/openfeature-node-provider/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @reflag/openfeature-node-provider
 
+## 1.2.6
+
+### Patch Changes
+
+- Updated dependencies [c69ee92]
+- Updated dependencies [df37a58]
+  - @reflag/node-sdk@1.8.1
+
 ## 1.2.5
 
 ### Patch Changes